Decode surface data

Correlates the significant clusters of an earlier vertex-wise analysis with a database of task-based fMRI and voxel-based morphometric statistical maps and associate them with relevant key words

Usage

decode_surf_data(surf_data, contrast = "positive", keywords_list = F)

Arguments

surf_data

a numeric vector with a length of 20484

contrast

A string object indicating whether to decode the positive or negative mask ('positive' or 'negative')

keywords_list

A boolean stating whether to return the full list of keywords in the imported neurosynth database

Value

A data.frame object listing the keywords and their Pearson's R values. If keywords_list is TRUE, a list containing the same data.frame object and the list of all the database keywords.

Details

The NiMARE python module is used for the imaging decoding and is imported via the reticulate package. The function needs the neurosynth database in the package's inst/extdata directory to be downloaded (~8 Mb) for the analysis. To maximise the speed of processing, keywords were excluded from the original database if they were: the name of an anatomical brain region, protocol-specific descriptive words that are not meaningful out of context (e.g., “studied”, “clips”, “engaged”), and broad categories that do not specifically point to a content (“disorder”, “psychological”, “cognitive”). A full list can be obtained with the keyword_list argument.

Examples

CTv = rbinom(20484, 1, 0.001) 
decode_surf_data(CTv, 'positive')
#> Checking for VertexWiseR system requirements ...
#> 
#> Converting and interpolating the surface data...
#> Correlating input image with images in the neurosynth database. This may take a while...
#>                     keyword      r
#> 449                    pain  0.037
#> 20                   adults  0.033
#> 57                autonomic  0.033
#> 111                decision  0.033
#> 194                 empathy  0.031
#> 198              engagement  0.030
#> 440            older adults  0.029
#> 26                    aging  0.028
#> 22                affective  0.026
#> 112         decision making  0.026
#> 379              motivation  0.026
#> 676                 valence  0.026
#> 63     behavioral responses  0.025
#> 343           matching task  0.025
#> 412    network connectivity  0.025
#> 439                   older  0.025
#> 624              strategies  0.025
#> 381                   motor  0.024
#> 34             anticipation  0.023
#> 219              extinction  0.023
#> 257                    gaze  0.023
#> 262           goal directed  0.023
#> 8                   actions  0.022
#> 280               imitation  0.022
#> 360              mesolimbic  0.022
#> 538               retrieval  0.022
#> 543                  reward  0.022
#> 332              linguistic  0.021
#> 342                matching  0.021
#> 405              navigation  0.021
#> 427                 noxious  0.021
#> 550             risk taking  0.021
#> 599        social cognition  0.021
#> 666             uncertainty  0.021
#> 675                updating  0.021
#> 117         default network  0.020
#> 187   emotional information  0.020
#> 237                feedback  0.020
#> 358          mental imagery  0.020
#> 609                 spatial  0.020
#> 21                   affect  0.019
#> 24                      age  0.019
#> 60                avoidance  0.019
#> 89             compensatory  0.019
#> 141            discriminate  0.019
#> 195                encoding  0.019
#> 251                 focused  0.019
#> 285               impulsive  0.019
#> 287               incentive  0.019
#> 355            memory tasks  0.019
#> 364          mild cognitive  0.019
#> 367           mirror neuron  0.019
#> 369                monetary  0.019
#> 420         neutral stimuli  0.019
#> 548                    rich  0.019
#> 551                   risky  0.019
#> 602     social interactions  0.019
#> 660                  traits  0.019
#> 19                adulthood  0.018
#> 29        alzheimer disease  0.018
#> 157     distributed network  0.018
#> 203         episodic memory  0.018
#> 216               expressed  0.018
#> 238                 feeling  0.018
#> 319                 learned  0.018
#> 353        memory retrieval  0.018
#> 455                patterns  0.018
#> 486             preparation  0.018
#> 592                 sighted  0.018
#> 654               time task  0.018
#> 655                     tom  0.018
#> 695                visually  0.018
#> 28                alzheimer  0.017
#> 69                    blind  0.017
#> 145          discriminative  0.017
#> 185               emotional  0.017
#> 234                    fear  0.017
#> 302           interpersonal  0.017
#> 336                  losses  0.017
#> 348                  memory  0.017
#> 359             mentalizing  0.017
#> 576              sensations  0.017
#> 625                strategy  0.017
#> 635               sustained  0.017
#> 643                  target  0.017
#> 710                   young  0.017
#> 711            young adults  0.017
#> 7        action observation  0.016
#> 43                attention  0.016
#> 55         autobiographical  0.016
#> 56  autobiographical memory  0.016
#> 77                  chinese  0.016
#> 202                episodic  0.016
#> 204                   error  0.016
#> 261                    goal  0.016
#> 371         monetary reward  0.016
#> 375              monitoring  0.016
#> 410       negative feedback  0.016
#> 474                planning  0.016
#> 480              prediction  0.016
#> 531     response inhibition  0.016
#> 534                    rest  0.016
#> 714          younger adults  0.016
#> 6                    action  0.015
#> 54          autism spectrum  0.015
#> 58                 aversive  0.015
#> 284                implicit  0.015
#> 385       motor performance  0.015
#> 438                     old  0.015
#> 448                   paced  0.015
#> 88             compensation  0.014
#> 114               decisions  0.014
#> 144     discrimination task  0.014
#> 245                fixation  0.014
#> 274        high functioning  0.014
#> 301           interoceptive  0.014
#> 365                mind tom  0.014
#> 380            motivational  0.014
#> 437                 oddball  0.014
#> 499                reaction  0.014
#> 546                 rewards  0.014
#> 559                salience  0.014
#> 561                 salient  0.014
#> 623               strategic  0.014
#> 712             young older  0.014
#> 713                 younger  0.014
#> 41                  arousal  0.013
#> 90                competing  0.013
#> 101             conflicting  0.013
#> 143          discrimination  0.013
#> 156                distress  0.013
#> 211              expectancy  0.013
#> 464              peripheral  0.013
#> 477                pleasant  0.013
#> 478                positive  0.013
#> 510             recognition  0.013
#> 555                   rules  0.013
#> 611     spatial information  0.013
#> 616      spectrum disorders  0.013
#> 645                 targets  0.013
#> 697           visuo spatial  0.013
#> 78                   choice  0.012
#> 115                 default  0.012
#> 181               effortful  0.012
#> 196      encoding retrieval  0.012
#> 199                engaging  0.012
#> 205                  errors  0.012
#> 214               expertise  0.012
#> 273              heart rate  0.012
#> 320                learning  0.012
#> 345                     mci  0.012
#> 366                  mirror  0.012
#> 398                 musical  0.012
#> 541                 reverse  0.012
#> 644        target detection  0.012
#> 665               uncertain  0.012
#> 31                 amnestic  0.011
#> 42                      asd  0.011
#> 61                awareness  0.011
#> 183                 emotion  0.011
#> 188     emotional responses  0.011
#> 200                 english  0.011
#> 233             familiarity  0.011
#> 272                   heart  0.011
#> 349         memory encoding  0.011
#> 384           motor imagery  0.011
#> 454                 pattern  0.011
#> 500           reaction time  0.011
#> 507             reappraisal  0.011
#> 554                    rule  0.011
#> 580                 sensory  0.011
#> 631              subjective  0.011
#> 2                  abstract  0.010
#> 79                  choices  0.010
#> 104   congruent incongruent  0.010
#> 113           decision task  0.010
#> 122                dementia  0.010
#> 173                    dual  0.010
#> 190       emotional valence  0.010
#> 193                empathic  0.010
#> 212             expectation  0.010
#> 255                gambling  0.010
#> 328        lexical decision  0.010
#> 377                   moral  0.010
#> 417                 neutral  0.010
#> 444             orientation  0.010
#> 450                 painful  0.010
#> 487             preparatory  0.010
#> 533          response times  0.010
#> 537               retention  0.010
#> 544     reward anticipation  0.010
#> 652                  threat  0.010
#> 674                upcoming  0.010
#> 684                  visual  0.010
#> 686         visual auditory  0.010
#> 699            visuospatial  0.010
#> 53                   autism  0.009
#> 119                   delay  0.009
#> 180                  effort  0.009
#> 278                 imagine  0.009
#> 291                 inhibit  0.009
#> 296                insights  0.009
#> 388          motor response  0.009
#> 403            naturalistic  0.009
#> 429                 numbers  0.009
#> 482             predictions  0.009
#> 512        recognition task  0.009
#> 598                  social  0.009
#> 615       spectrum disorder  0.009
#> 622             stop signal  0.009
#> 690       visual perception  0.009
#> 10                 adaptive  0.008
#> 70                   bodily  0.008
#> 130                  detect  0.008
#> 166            dopaminergic  0.008
#> 171                    drug  0.008
#> 208       executive control  0.008
#> 288         incentive delay  0.008
#> 318                   learn  0.008
#> 382           motor control  0.008
#> 411              negativity  0.008
#> 446               orienting  0.008
#> 485            preferential  0.008
#> 504                    read  0.008
#> 516            recollection  0.008
#> 519              regulatory  0.008
#> 641                 tactile  0.008
#> 642                 tapping  0.008
#> 661                transfer  0.008
#> 672              unpleasant  0.008
#> 692         visual stimulus  0.008
#> 14              adolescence  0.007
#> 83      cognitive emotional  0.007
#> 84     cognitive impairment  0.007
#> 91              competition  0.007
#> 103               congruent  0.007
#> 184      emotion regulation  0.007
#> 186         emotional faces  0.007
#> 207               executive  0.007
#> 235                 fearful  0.007
#> 243        finger movements  0.007
#> 247                flexible  0.007
#> 254  frontoparietal network  0.007
#> 260                gestures  0.007
#> 266                  handed  0.007
#> 279                imagined  0.007
#> 292              inhibition  0.007
#> 321           learning task  0.007
#> 372                   money  0.007
#> 459              perceiving  0.007
#> 501          reaction times  0.007
#> 523  reinforcement learning  0.007
#> 560        salience network  0.007
#> 568     selective attention  0.007
#> 593             signal task  0.007
#> 683                  vision  0.007
#> 102              congruency  0.006
#> 142           discriminated  0.006
#> 155             distractors  0.006
#> 177                  eating  0.006
#> 182                 elderly  0.006
#> 210     executive functions  0.006
#> 217              expression  0.006
#> 290            index finger  0.006
#> 386               motor pre  0.006
#> 453         passive viewing  0.006
#> 490              production  0.006
#> 498              punishment  0.006
#> 532           response time  0.006
#> 572         semantic memory  0.006
#> 586               serotonin  0.006
#> 619       speech production  0.006
#> 639                symbolic  0.006
#> 693           visual stream  0.006
#> 698              visuomotor  0.006
#> 709          working memory  0.006
#> 1                 abilities  0.005
#> 12                 addition  0.005
#> 18                    adult  0.005
#> 23                  affects  0.005
#> 35             anticipatory  0.005
#> 109         control network  0.005
#> 132          detection task  0.005
#> 170                 driving  0.005
#> 209      executive function  0.005
#> 215                explicit  0.005
#> 224           eye movements  0.005
#> 289             incongruent  0.005
#> 356             memory test  0.005
#> 404                  nature  0.005
#> 430               numerical  0.005
#> 462              perceptual  0.005
#> 494         psychopathology  0.005
#> 508               reasoning  0.005
#> 521               rehearsal  0.005
#> 522           reinforcement  0.005
#> 545               rewarding  0.005
#> 549                    risk  0.005
#> 603                socially  0.005
#> 646                   taste  0.005
#> 664                    true  0.005
#> 45        attention network  0.004
#> 100                conflict  0.004
#> 110                 craving  0.004
#> 162          domain general  0.004
#> 189       emotional stimuli  0.004
#> 268                   hands  0.004
#> 282              impairment  0.004
#> 308           judgment task  0.004
#> 331                  limbic  0.004
#> 357                     men  0.004
#> 368                mnemonic  0.004
#> 370      monetary incentive  0.004
#> 392               movements  0.004
#> 407         negative affect  0.004
#> 426                 novelty  0.004
#> 428                  number  0.004
#> 442                    oral  0.004
#> 447            orthographic  0.004
#> 481        prediction error  0.004
#> 509                  recall  0.004
#> 526                remember  0.004
#> 528             remembering  0.004
#> 566                 seeking  0.004
#> 575               sensation  0.004
#> 591              short term  0.004
#> 613                 speaker  0.004
#> 657                   tones  0.004
#> 64                   belief  0.003
#> 131               detection  0.003
#> 165                dopamine  0.003
#> 232                familiar  0.003
#> 239                feelings  0.003
#> 286             impulsivity  0.003
#> 309               judgments  0.003
#> 335                    loss  0.003
#> 374               monitored  0.003
#> 387          motor premotor  0.003
#> 390              motor task  0.003
#> 406                negative  0.003
#> 425                   novel  0.003
#> 443                   order  0.003
#> 458               perceived  0.003
#> 463        performance task  0.003
#> 502               reactions  0.003
#> 514              recognized  0.003
#> 634             subtraction  0.003
#> 685        visual attention  0.003
#> 703                   women  0.003
#> 33                    angry  0.002
#> 139              disability  0.002
#> 154              distractor  0.002
#> 163               dominance  0.002
#> 191             emotionally  0.002
#> 228                   faces  0.002
#> 231      facial expressions  0.002
#> 265          hand movements  0.002
#> 297            intelligence  0.002
#> 350             memory load  0.002
#> 419        neutral pictures  0.002
#> 421             nociceptive  0.002
#> 457                perceive  0.002
#> 581     sensory information  0.002
#> 588                  sexual  0.002
#> 604                 solving  0.002
#> 633               substance  0.002
#> 651             theory mind  0.002
#> 688      visual information  0.002
#> 5                       act  0.001
#> 13                     adhd  0.001
#> 17                  adopted  0.001
#> 25                     aged  0.001
#> 47              attentional  0.001
#> 65                  beliefs  0.001
#> 76     children adolescents  0.001
#> 80             chronic pain  0.001
#> 85                    color  0.001
#> 218             expressions  0.001
#> 244          finger tapping  0.001
#> 256                    game  0.001
#> 293              inhibitory  0.001
#> 307                judgment  0.001
#> 315              laterality  0.001
#> 334               listening  0.001
#> 352        memory processes  0.001
#> 393                  moving  0.001
#> 422                    nogo  0.001
#> 431                  object  0.001
#> 433                 objects  0.001
#> 471                pictures  0.001
#> 511      recognition memory  0.001
#> 518              regulation  0.001
#> 553                      rt  0.001
#> 569                    self  0.001
#> 587                     sex  0.001
#> 594                  silent  0.001
#> 636     sustained attention  0.001
#> 658                   touch  0.001
#> 668              unexpected  0.001
#> 696                   visuo  0.001
#> 11                addiction  0.000
#> 32                    anger  0.000
#> 44        attention deficit  0.000
#> 52          auditory visual  0.000
#> 59                    avoid  0.000
#> 106           consciousness  0.000
#> 116            default mode  0.000
#> 140                discrete  0.000
#> 169                   drive  0.000
#> 197              endogenous  0.000
#> 222              eye fields  0.000
#> 271                 hearing  0.000
#> 299             intentional  0.000
#> 310               knowledge  0.000
#> 324                  letter  0.000
#> 351      memory performance  0.000
#> 373                 monitor  0.000
#> 394      multiple sclerosis  0.000
#> 465             personality  0.000
#> 469             photographs  0.000
#> 552                rotation  0.000
#> 601      social interaction  0.000
#> 605           somatosensory  0.000
#> 638               switching  0.000
#> 681                   verbs  0.000
#> 118   deficit hyperactivity -0.001
#> 135           developmental -0.001
#> 220                     eye -0.001
#> 226                    face -0.001
#> 230       facial expression -0.001
#> 246             flexibility -0.001
#> 249                 fluency -0.001
#> 263                   goals -0.001
#> 276  hyperactivity disorder -0.001
#> 341                   match -0.001
#> 391                movement -0.001
#> 395            multisensory -0.001
#> 414       neurodegenerative -0.001
#> 418           neutral faces -0.001
#> 484             preferences -0.001
#> 491             pseudowords -0.001
#> 515             recognizing -0.001
#> 600        social cognitive -0.001
#> 630             stroop task -0.001
#> 632       subsequent memory -0.001
#> 650             term memory -0.001
#> 659                   trait -0.001
#> 673               unrelated -0.001
#> 700                   vocal -0.001
#> 86            communication -0.002
#> 105               conscious -0.002
#> 128             deprivation -0.002
#> 149           disorder adhd -0.002
#> 159               disturbed -0.002
#> 225                    eyes -0.002
#> 305                   judge -0.002
#> 347                memories -0.002
#> 361                     met -0.002
#> 383          motor function -0.002
#> 467             perspective -0.002
#> 473                   place -0.002
#> 479           posttraumatic -0.002
#> 506                 reality -0.002
#> 529              repetition -0.002
#> 565                  search -0.002
#> 577                   sense -0.002
#> 579    sensorimotor network -0.002
#> 589                   shape -0.002
#> 590                  shapes -0.002
#> 637                  switch -0.002
#> 656                    tone -0.002
#> 667              understand -0.002
#> 682              violations -0.002
#> 707        word recognition -0.002
#> 3                     abuse -0.003
#> 16              adolescents -0.003
#> 46           attention task -0.003
#> 81        cognitive control -0.003
#> 82       cognitive deficits -0.003
#> 92            comprehension -0.003
#> 95            concentration -0.003
#> 120                 delayed -0.003
#> 167        dorsal attention -0.003
#> 168                     dot -0.003
#> 221               eye field -0.003
#> 223            eye movement -0.003
#> 229                  facial -0.003
#> 236           fearful faces -0.003
#> 241                 females -0.003
#> 294      inhibitory control -0.003
#> 295                 insight -0.003
#> 300              intentions -0.003
#> 314               languages -0.003
#> 339                    male -0.003
#> 397                   music -0.003
#> 452       parkinson disease -0.003
#> 520          rehabilitation -0.003
#> 540                reversal -0.003
#> 564               sclerosis -0.003
#> 573            semantically -0.003
#>  [ reached 'max' / getOption("max.print") -- omitted 215 rows ]