Surface to atlas

Returns the mean or sum of vertex-wise surface data for each ROI of a selected atlas

Usage

surf_to_atlas(surf_data, atlas, mode = "mean")

Arguments

surf_data

A matrix object containing the surface data in fsaverage5 (20484 vertices), fsaverage6 (81924 vertices) or hippocampal (14524 vertices) space. See also Hipvextract() or SURFvextract() output format.

atlas

A numeric integer object corresponding to the atlas of interest. 1=aparc, 2=Destrieux-148, 3=Glasser-360, 4=Schaefer-100, 5=Schaefer-200, 6=Schaefer-400. For hippocampal surface, the bigbrain hippocampal atlas is used by default and ignores the option.

mode

A string indicating whether to extract the sum ('sum') or the average ('mean') of the ROI vertices values. Default is 'mean'.

Value

A matrix object with ROI as column and corresponding average vertex-wise values as row

Details

The function currently works with the aparc/Desikan-Killiany-70, Destrieux-148, Glasser-360, Schaefer-100, Schaefer-200, Schaefer-400 atlases. ROI to vertex mapping data were obtained from the enigmatoolbox ; data for Destrieux came from nilearn.datasets.fetch_atlas_surf_destrieux

For hippocampal data, the function currently works with the "bigbrain" atlas integrated in Hippunfold. See also DeKraker et al., 2020.

See also

Examples

CTv = runif(20484,min=0, max=100)
surf_to_atlas(CTv, 1)
#>  [1] 49.92331 53.49355 47.22426 52.77122 51.15864 52.09127 50.01991 51.31866
#>  [9] 50.28869 51.50557 50.09961 51.72056 53.05888 50.59699 47.78444 50.90100
#> [17] 50.74526 49.69928 44.11279 50.14934 46.51242 47.83549 48.90736 50.63215
#> [25] 50.04727 45.26273 51.54100 50.44109 48.27340 49.56037 50.57570 65.78174
#> [33] 50.54350 47.99889 48.97449 49.64571 50.78299 45.35281 51.13575 49.53441
#> [41] 43.19311 51.48077 49.23856 53.24017 48.55781 46.45040 48.17006 46.94278
#> [49] 48.08163 49.23326 50.26257 46.32722 47.28163 50.78247 51.82155 49.48975
#> [57] 52.67509 48.32513 49.40488 51.88156 53.15854 49.17489 49.86591 50.19587
#> [65] 50.51372 49.40957 54.74496 46.72882 56.40791 50.86727