Vertex-wise analysis

Fits a linear or linear mixed model with the cortical or hippocampal surface data as the predicted outcome, and returns cluster-thresholded (Random field theory) t-stat map selected contrast.

Usage

vertex_analysis(
  model,
  contrast,
  random,
  surf_data,
  p = 0.05,
  atlas = 1,
  smooth_FWHM
)

Arguments

model

An N X V data.frame object containing N rows for each subject and V columns for each predictor included in the model. This data.frame should not include the random effects variable.

contrast

A numeric vector or object containing the values of the predictor of interest. The cluster-thresholded t-stat maps will be estimated only for this predictor

random

An object containing the values of the random variable (optional)

surf_data

A matrix object containing the surface data, see SURFvextract() or HIPvextract() output format.

p

A numeric object specifying the p-value to threshold the results (Default is 0.05)

atlas

A numeric integer object corresponding to the atlas of interest. 1=Desikan, 2=Schaefer-100, 3=Schaefer-200, 4=Glasser-360, 5=Destrieux-148.

smooth_FWHM

A numeric vector object specifying the desired smoothing width in mm

Value

A list object containing the cluster level results, thresholded t-stat map, and positive, negative and bidirectional cluster maps.

Details

The function imports and adapts the brainstat python library.

Output definitions:

nverts: number of vertices in the cluster

P: p-value of the cluster

X, Y and Z: MNI coordinates of the vertex with the highest t-statistic in the cluster.

tstat: t statistic of the vertex with the highest t-statistic in the cluster

region: the region this highest -statistic vertex is located in, as determined/labelled by the selected atlas

Examples

demodata = readRDS(system.file('demo_data/SPRENG_behdata_site1.rds', 
package = 'VertexWiseR'))[1:100,]
CTv = readRDS(file = url(paste0("https://github.com",
"/CogBrainHealthLab/VertexWiseR/blob/main/inst/demo_data/",
"SPRENG_CTv_site1.rds?raw=TRUE")))[1:100,]

vertexwise_model=vertex_analysis(model=demodata[,c(2,7)], 
contrast=demodata[,7], surf_data = CTv, atlas=1,p = 0.05)
#> The binary variable 'gender' will be recoded with F=0 and M=1 for the analysis
#> smooth_FWHM argument was not given. surf_data will not be smoothed here.
print(vertexwise_model$cluster_level_results)
#> $`Positive contrast`
#>   clusid nverts      P     X    Y     Z tstat                  region
#> 1      1     64 <0.001 -28.0 -1.7 -43.5  5.25             lh-fusiform
#> 2      2     35  0.002 -14.6 11.7 -24.6  5.00 lh-lateralorbitofrontal
#> 3      3     35  0.008  27.9  8.8 -44.2  4.48         rh-temporalpole
#> 
#> $`Negative contrast`
#>    clusid nverts      P     X     Y     Z tstat                 region
#> 1       1   3532 <0.001  46.2   4.8 -18.2 -7.57    rh-superiortemporal
#> 2       2   3489 <0.001 -35.0 -21.5  68.8 -8.23          lh-precentral
#> 3       3    298 <0.001 -19.0 -64.4   1.8 -5.06             lh-lingual
#> 4       4    145 <0.001  32.5 -44.6  -9.3 -6.14     rh-parahippocampal
#> 5       5     48  0.001 -48.3 -27.9  59.9 -5.16         lh-postcentral
#> 6       6     30  0.001  -8.1  64.7 -13.0 -3.66         lh-frontalpole
#> 7       7     40  0.004   4.2 -10.9  67.2 -4.09         rh-paracentral
#> 8       8     34  0.004 -32.9 -41.1 -12.1 -3.90     lh-parahippocampal
#> 9       9     30  0.013 -43.7 -48.3 -16.7 -3.36            lh-fusiform
#> 10     10     28   0.02  12.3  20.5 -17.4 -4.24 rh-medialorbitofrontal
#> 11     11     23  0.033   2.7  48.4 -18.6 -4.74 rh-medialorbitofrontal
#>