Decode surface data

Correlates the significant clusters of an earlier vertex-wise analysis with a database of task-based fMRI and voxel-based morphometric statistical maps and associate them with relevant key words

Usage

decode_surf_data(surf_data, contrast = "positive")

Arguments

surf_data

a numeric vector with a length of 20484

contrast

A string object indicating whether to decode the positive or negative mask ('positive' or 'negative')

Value

A data.frame object listing the keywords and their Pearson's R values

Details

The NiMARE python module is used for the imaging decoding and is imported via the reticulate package. The function also downloads the neurosynth database in the package's inst/extdata direcotry (~8 Mb) for the analysis.

Examples

CTv = rbinom(20484, 1, 0.001) 
decode_surf_data(CTv, 'positive')
#> Correlating input image with images in the neurosynth database. This may take a while
#>                     keyword      r
#> 195                encoding  0.052
#> 538               retrieval  0.051
#> 311                language  0.046
#> 510             recognition  0.044
#> 202                episodic  0.043
#> 381                   motor  0.041
#> 571                semantic  0.041
#> 509                  recall  0.039
#> 348                  memory  0.038
#> 704                    word  0.036
#> 55         autobiographical  0.035
#> 198              engagement  0.035
#> 424                   nouns  0.035
#> 232                familiar  0.032
#> 425                   novel  0.032
#> 507             reappraisal  0.032
#> 620                  spoken  0.032
#> 648                temporal  0.032
#> 681                   verbs  0.032
#> 347                memories  0.031
#> 391                movement  0.031
#> 449                    pain  0.031
#> 200                 english  0.030
#> 713                 younger  0.030
#> 708                   words  0.029
#> 194                 empathy  0.028
#> 203         episodic memory  0.028
#> 233             familiarity  0.028
#> 314               languages  0.028
#> 455                patterns  0.027
#> 6                    action  0.026
#> 173                    dual  0.026
#> 332              linguistic  0.026
#> 417                 neutral  0.026
#> 521               rehearsal  0.026
#> 308           judgment task  0.025
#> 384           motor imagery  0.025
#> 438                     old  0.024
#> 467             perspective  0.024
#> 56  autobiographical memory  0.023
#> 92            comprehension  0.023
#> 98               conceptual  0.023
#> 113           decision task  0.023
#> 279                imagined  0.023
#> 284                implicit  0.023
#> 306                  judged  0.023
#> 313        language network  0.023
#> 327                 lexical  0.023
#> 431                  object  0.023
#> 490              production  0.023
#> 516            recollection  0.023
#> 60                avoidance  0.022
#> 196      encoding retrieval  0.022
#> 242                  finger  0.022
#> 290            index finger  0.022
#> 328        lexical decision  0.022
#> 439                   older  0.022
#> 440            older adults  0.022
#> 471                pictures  0.022
#> 506                 reality  0.022
#> 528             remembering  0.022
#> 529              repetition  0.022
#> 572         semantic memory  0.022
#> 669              unfamiliar  0.022
#> 8                   actions  0.021
#> 77                  chinese  0.021
#> 234                    fear  0.021
#> 385       motor performance  0.021
#> 447            orthographic  0.021
#> 537               retention  0.021
#> 583                sentence  0.021
#> 613                 speaker  0.021
#> 682              violations  0.021
#> 214               expertise  0.020
#> 280               imitation  0.020
#> 333                listened  0.020
#> 349         memory encoding  0.020
#> 392               movements  0.020
#> 548                    rich  0.020
#> 619       speech production  0.020
#> 677                    verb  0.020
#> 20                   adults  0.019
#> 157     distributed network  0.019
#> 184      emotion regulation  0.019
#> 219              extinction  0.019
#> 366                  mirror  0.019
#> 367           mirror neuron  0.019
#> 459              perceiving  0.019
#> 470                 picture  0.019
#> 578            sensorimotor  0.019
#> 602     social interactions  0.019
#> 714          younger adults  0.019
#> 243        finger movements  0.018
#> 260                gestures  0.018
#> 353        memory retrieval  0.018
#> 355            memory tasks  0.018
#> 448                   paced  0.018
#> 454                 pattern  0.018
#> 502               reactions  0.018
#> 512        recognition task  0.018
#> 518              regulation  0.018
#> 605           somatosensory  0.018
#> 26                    aging  0.017
#> 58                 aversive  0.017
#> 187   emotional information  0.017
#> 209      executive function  0.017
#> 266                  handed  0.017
#> 302           interpersonal  0.017
#> 312  language comprehension  0.017
#> 320                learning  0.017
#> 468            phonological  0.017
#> 576              sensations  0.017
#> 678                  verbal  0.017
#> 705               word form  0.017
#> 7        action observation  0.016
#> 22                affective  0.016
#> 88             compensation  0.016
#> 193                empathic  0.016
#> 358          mental imagery  0.016
#> 427                 noxious  0.016
#> 450                 painful  0.016
#> 491             pseudowords  0.016
#> 591              short term  0.016
#> 616      spectrum disorders  0.016
#> 642                 tapping  0.016
#> 655                     tom  0.016
#> 667              understand  0.016
#> 694             visual word  0.016
#> 249                 fluency  0.015
#> 278                 imagine  0.015
#> 330                    limb  0.015
#> 398                 musical  0.015
#> 485            preferential  0.015
#> 99               confidence  0.014
#> 238                 feeling  0.014
#> 244          finger tapping  0.014
#> 307                judgment  0.014
#> 382           motor control  0.014
#> 387          motor premotor  0.014
#> 463        performance task  0.014
#> 486             preparation  0.014
#> 539               retrieved  0.014
#> 564               sclerosis  0.014
#> 695                visually  0.014
#> 57                autonomic  0.013
#> 352        memory processes  0.013
#> 433                 objects  0.013
#> 526                remember  0.013
#> 527              remembered  0.013
#> 584  sentence comprehension  0.013
#> 700                   vocal  0.013
#> 10                 adaptive  0.012
#> 21                   affect  0.012
#> 31                 amnestic  0.012
#> 142           discriminated  0.012
#> 183                 emotion  0.012
#> 264                    hand  0.012
#> 317             lateralized  0.012
#> 365                mind tom  0.012
#> 511      recognition memory  0.012
#> 585               sentences  0.012
#> 617                  speech  0.012
#> 644        target detection  0.012
#> 710                   young  0.012
#> 84     cognitive impairment  0.011
#> 164                dominant  0.011
#> 215                explicit  0.011
#> 218             expressions  0.011
#> 237                feedback  0.011
#> 248                flexibly  0.011
#> 357                     men  0.011
#> 364          mild cognitive  0.011
#> 517                regulate  0.011
#> 594                  silent  0.011
#> 598                  social  0.011
#> 653             threatening  0.011
#> 706              word pairs  0.011
#> 29        alzheimer disease  0.010
#> 32                    anger  0.010
#> 54          autism spectrum  0.010
#> 117         default network  0.010
#> 145          discriminative  0.010
#> 303                      iq  0.010
#> 309               judgments  0.010
#> 400                   names  0.010
#> 580                 sensory  0.010
#> 635               sustained  0.010
#> 641                 tactile  0.010
#> 668              unexpected  0.010
#> 673               unrelated  0.010
#> 693           visual stream  0.010
#> 712             young older  0.010
#> 69                    blind  0.009
#> 156                distress  0.009
#> 191             emotionally  0.009
#> 315              laterality  0.009
#> 360              mesolimbic  0.009
#> 390              motor task  0.009
#> 394      multiple sclerosis  0.009
#> 552                rotation  0.009
#> 631              subjective  0.009
#> 640               syntactic  0.009
#> 679          verbal fluency  0.009
#> 132          detection task  0.008
#> 135           developmental  0.008
#> 259                  german  0.008
#> 388          motor response  0.008
#> 474                planning  0.008
#> 615       spectrum disorder  0.008
#> 625                strategy  0.008
#> 632       subsequent memory  0.008
#> 650             term memory  0.008
#> 664                    true  0.008
#> 698              visuomotor  0.008
#> 707        word recognition  0.008
#> 24                      age  0.007
#> 28                alzheimer  0.007
#> 89             compensatory  0.007
#> 101             conflicting  0.007
#> 118   deficit hyperactivity  0.007
#> 199                engaging  0.007
#> 272                   heart  0.007
#> 276  hyperactivity disorder  0.007
#> 368                mnemonic  0.007
#> 375              monitoring  0.007
#> 401                  naming  0.007
#> 405              navigation  0.007
#> 419        neutral pictures  0.007
#> 426                 novelty  0.007
#> 437                 oddball  0.007
#> 461              perception  0.007
#> 505                 readers  0.007
#> 514              recognized  0.007
#> 575               sensation  0.007
#> 592                 sighted  0.007
#> 599        social cognition  0.007
#> 690       visual perception  0.007
#> 3                     abuse  0.006
#> 38                  aphasia  0.006
#> 63     behavioral responses  0.006
#> 114               decisions  0.006
#> 158            disturbances  0.006
#> 182                 elderly  0.006
#> 211              expectancy  0.006
#> 239                feelings  0.006
#> 247                flexible  0.006
#> 319                 learned  0.006
#> 342                matching  0.006
#> 345                     mci  0.006
#> 359             mentalizing  0.006
#> 396                  muscle  0.006
#> 412    network connectivity  0.006
#> 420         neutral stimuli  0.006
#> 478                positive  0.006
#> 520          rehabilitation  0.006
#> 544     reward anticipation  0.006
#> 555                   rules  0.006
#> 42                      asd  0.005
#> 44        attention deficit  0.005
#> 53                   autism  0.005
#> 83      cognitive emotional  0.005
#> 87               compensate  0.005
#> 96                  concept  0.005
#> 192                emotions  0.005
#> 207               executive  0.005
#> 216               expressed  0.005
#> 253                    foot  0.005
#> 255                gambling  0.005
#> 256                    game  0.005
#> 257                    gaze  0.005
#> 262           goal directed  0.005
#> 271                 hearing  0.005
#> 275           hyperactivity  0.005
#> 299             intentional  0.005
#> 379              motivation  0.005
#> 386               motor pre  0.005
#> 422                    nogo  0.005
#> 443                   order  0.005
#> 513               recognize  0.005
#> 533          response times  0.005
#> 547                  rhythm  0.005
#> 554                    rule  0.005
#> 582           sensory motor  0.005
#> 595                   sleep  0.005
#> 624              strategies  0.005
#> 65                  beliefs  0.004
#> 80             chronic pain  0.004
#> 321           learning task  0.004
#> 324                  letter  0.004
#> 343           matching task  0.004
#> 399               musicians  0.004
#> 407         negative affect  0.004
#> 444             orientation  0.004
#> 469             photographs  0.004
#> 515             recognizing  0.004
#> 573            semantically  0.004
#> 590                  shapes  0.004
#> 623               strategic  0.004
#> 660                  traits  0.004
#> 676                 valence  0.004
#> 696                   visuo  0.004
#> 697           visuo spatial  0.004
#> 23                  affects  0.003
#> 30                ambiguous  0.003
#> 91              competition  0.003
#> 119                   delay  0.003
#> 148                 disgust  0.003
#> 181               effortful  0.003
#> 212             expectation  0.003
#> 217              expression  0.003
#> 268                   hands  0.003
#> 305                   judge  0.003
#> 350             memory load  0.003
#> 409       negative emotions  0.003
#> 482             predictions  0.003
#> 530  repetition suppression  0.003
#> 534                    rest  0.003
#> 541                 reverse  0.003
#> 601      social interaction  0.003
#> 614                speaking  0.003
#> 638               switching  0.003
#> 646                   taste  0.003
#> 652                  threat  0.003
#> 680          verbal working  0.003
#> 2                  abstract  0.002
#> 34             anticipation  0.002
#> 41                  arousal  0.002
#> 66               bilinguals  0.002
#> 76     children adolescents  0.002
#> 82       cognitive deficits  0.002
#> 106           consciousness  0.002
#> 110                 craving  0.002
#> 159               disturbed  0.002
#> 162          domain general  0.002
#> 241                 females  0.002
#> 251                 focused  0.002
#> 252                    food  0.002
#> 273              heart rate  0.002
#> 331                  limbic  0.002
#> 339                    male  0.002
#> 341                   match  0.002
#> 383          motor function  0.002
#> 397                   music  0.002
#> 410       negative feedback  0.002
#> 421             nociceptive  0.002
#> 574               semantics  0.002
#> 581     sensory information  0.002
#> 651             theory mind  0.002
#> 666             uncertainty  0.002
#> 703                   women  0.002
#> 17                  adopted  0.001
#> 19                adulthood  0.001
#> 109         control network  0.001
#> 139              disability  0.001
#> 143          discrimination  0.001
#> 250                   focus  0.001
#> 274        high functioning  0.001
#> 291                 inhibit  0.001
#> 318                   learn  0.001
#> 334               listening  0.001
#> 344              maturation  0.001
#> 373                 monitor  0.001
#> 458               perceived  0.001
#> 472                   pitch  0.001
#> 498              punishment  0.001
#> 504                    read  0.001
#> 560        salience network  0.001
#> 618       speech perception  0.001
#> 658                   touch  0.001
#> 670              unilateral  0.001
#> 675                updating  0.001
#> 685        visual attention  0.001
#> 1                 abilities  0.000
#> 11                addiction  0.000
#> 33                    angry  0.000
#> 64                   belief  0.000
#> 140                discrete  0.000
#> 153             distraction  0.000
#> 154              distractor  0.000
#> 188     emotional responses  0.000
#> 231      facial expressions  0.000
#> 285               impulsive  0.000
#> 289             incongruent  0.000
#> 304                isolated  0.000
#> 310               knowledge  0.000
#> 351      memory performance  0.000
#> 377                   moral  0.000
#> 403            naturalistic  0.000
#> 453         passive viewing  0.000
#> 519              regulatory  0.000
#> 531     response inhibition  0.000
#> 546                 rewards  0.000
#> 551                   risky  0.000
#> 654               time task  0.000
#> 672              unpleasant  0.000
#> 5                       act -0.001
#> 12                 addition -0.001
#> 46           attention task -0.001
#> 79                  choices -0.001
#> 97                 concepts -0.001
#> 107           consolidation -0.001
#> 128             deprivation -0.001
#> 147              disease pd -0.001
#> 170                 driving -0.001
#> 174                dyslexia -0.001
#> 204                   error -0.001
#> 228                   faces -0.001
#> 294      inhibitory control -0.001
#> 301           interoceptive -0.001
#> 316          lateralization -0.001
#> 380            motivational -0.001
#> 402                 natural -0.001
#> 428                  number -0.001
#> 477                pleasant -0.001
#> 496               psychotic -0.001
#> 508               reasoning -0.001
#> 523  reinforcement learning -0.001
#> 563           schizophrenic -0.001
#> 568     selective attention -0.001
#> 593             signal task -0.001
#> 600        social cognitive -0.001
#> 634             subtraction -0.001
#> 663               traumatic -0.001
#> 699            visuospatial -0.001
#> 711            young adults -0.001
#> 52          auditory visual -0.002
#> 115                 default -0.002
#> 120                 delayed -0.002
#> 141            discriminate -0.002
#> 230       facial expression -0.002
#> 245                fixation -0.002
#> 265          hand movements -0.002
#> 281                impaired -0.002
#> 282              impairment -0.002
#> 296                insights -0.002
#> 356             memory test -0.002
#> 418           neutral faces -0.002
#> 456                      pd -0.002
#> 480              prediction -0.002
#> 525           relationships -0.002
#> 611     spatial information -0.002
#> 636     sustained attention -0.002
#> 637                  switch -0.002
#> 684                  visual -0.002
#> 75                 children -0.003
#> 86            communication -0.003
#> 100                conflict -0.003
#> 131               detection -0.003
#> 180                  effort -0.003
#> 185               emotional -0.003
#> 213            expectations -0.003
#> 261                    goal -0.003
#> 283             impairments -0.003
#> 336                  losses -0.003
#> 354             memory task -0.003
#> 395            multisensory -0.003
#> 416        neuropsychiatric -0.003
#> 442                    oral -0.003
#> 466      personality traits -0.003
#> 487             preparatory -0.003
#> 569                    self -0.003
#> 579    sensorimotor network -0.003
#> 603                socially -0.003
#> 656                    tone -0.003
#> 40                      arm -0.004
#> 70                   bodily -0.004
#> 111                decision -0.004
#> 166            dopaminergic -0.004
#> 167        dorsal attention -0.004
#> 178                economic -0.004
#> 208       executive control -0.004
#> 226                    face -0.004
#> 269                   happy -0.004
#> 445                oriented -0.004
#> 484             preferences -0.004
#> 499                reaction -0.004
#> 522           reinforcement -0.004
#> 543                  reward -0.004
#> 558                     sad -0.004
#> 577                   sense -0.004
#> 686         visual auditory -0.004
#> 688      visual information -0.004
#> 73           categorization -0.005
#> 74                childhood -0.005
#> 103               congruent -0.005
#> 121               demanding -0.005
#> 169                   drive -0.005
#> 176                    easy -0.005
#> 201                epilepsy -0.005
#> 229                  facial -0.005
#> 295                 insight -0.005
#> 298               intention -0.005
#> 300              intentions -0.005
#> 414       neurodegenerative -0.005
#> 457                perceive -0.005
#> 460                 percept -0.005
#> 464              peripheral -0.005
#> 550             risk taking -0.005
#> 647                      td -0.005
#> 687            visual field -0.005
#> 13                     adhd -0.006
#> 18                    adult -0.006
#> 62   behavioral performance -0.006
#> 67                  bipolar -0.006
#> 90                competing -0.006
#> 134             development -0.006
#> 149           disorder adhd -0.006
#> 155             distractors -0.006
#> 163               dominance -0.006
#> 227        face recognition -0.006
#> 267              handedness -0.006
#> 287               incentive -0.006
#> 293              inhibitory -0.006
#> 374               monitored -0.006
#> 406                negative -0.006
#> 432      object recognition -0.006
#> 479           posttraumatic -0.006
#> 488                   prime -0.006
#> 501          reaction times -0.006
#> 532           response time -0.006
#> 542                reversed -0.006
#> 545               rewarding -0.006
#> 559                salience -0.006
#> 587                     sex -0.006
#> 597                 smoking -0.006
#> 622             stop signal -0.006
#> 671             unmedicated -0.006
#> 691          visual stimuli -0.006
#> 692         visual stimulus -0.006
#> 27                  alcohol -0.007
#> 71                     body -0.007
#> 122                dementia -0.007
#> 146              disease ad -0.007
#> 152           disorder ptsd -0.007
#> 225                    eyes -0.007
#> 263                   goals -0.007
#> 288         incentive delay -0.007
#> 340                   males -0.007
#> 371         monetary reward -0.007
#> 372                   money -0.007
#> 411              negativity -0.007
#> 415      neurodevelopmental -0.007
#> 430               numerical -0.007
#> 441               olfactory -0.007
#> 462              perceptual -0.007
#> 473                   place -0.007
#> 500           reaction time -0.007
#> 503              reactivity -0.007
#> 565                  search -0.007
#> 567               selective -0.007
#> 645                 targets -0.007
#> 661                transfer -0.007
#> 683                  vision -0.007
#> 45        attention network -0.008
#> 61                awareness -0.008
#> 104   congruent incongruent -0.008
#> 116            default mode -0.008
#> 205                  errors -0.008
#> 206               evolution -0.008
#> 325                 letters -0.008
#> 335                    loss -0.008
#> 369                monetary -0.008
#> 588                  sexual -0.008
#> 665               uncertain -0.008
#> 35             anticipatory -0.009
#> 72                 capacity -0.009
#> 102              congruency -0.009
#> 105               conscious -0.009
#> 254  frontoparietal network -0.009
#> 270             happy faces -0.009
#> 277                illusion -0.009
#> 297            intelligence -0.009
#> 329                   light -0.009
#> 452       parkinson disease -0.009
#> 493   psychiatric disorders -0.009
#> 497                    ptsd -0.009
#> 566                 seeking -0.009
#> 586               serotonin -0.009
#> 589                   shape -0.009
#> 627         stress disorder -0.009
#> 628                  stroke -0.009
#> 639                symbolic -0.009
#> 657                   tones -0.009
#> 709          working memory -0.009
#> 68         bipolar disorder -0.010
#> 123              dependence -0.010
#> 130                  detect -0.010
#> 144     discrimination task -0.010
#> 186         emotional faces -0.010
#> 197              endogenous -0.010
#> 223            eye movement -0.010
#> 246             flexibility -0.010
#> 258                  gender -0.010
#> 292              inhibition -0.010
#> 429                 numbers -0.010
#> 451               parkinson -0.010
#> 483              preference -0.010
#> 489                 priming -0.010
#> 553                      rt -0.010
#> 604                 solving -0.010
#> 606                   sound -0.010
#> 701                   voice -0.010
#> 715                   youth -0.010
#> 4                  acoustic -0.011
#> 16              adolescents -0.011
#> 95            concentration -0.011
#> 108             consumption -0.011
#> 129                   depth -0.011
#> 137               direction -0.011
#> 160                 divided -0.011
#> 168                     dot -0.011
#> 171                    drug -0.011
#> 190       emotional valence -0.011
#> 326            level signal -0.011
#> 393                  moving -0.011
#> 404                  nature -0.011
#> 481        prediction error -0.011
#> 49              audiovisual -0.012
#> 78                   choice -0.012
#> 94      compulsive disorder -0.012
#> 222              eye fields -0.012
#> 235                 fearful -0.012
#> 435    obsessive compulsive -0.012
#> 475                    play -0.012
#> 540                reversal -0.012
#> 608                    span -0.012
#> 643                  target -0.012
#> 649              temporally -0.012
#> 25                     aged -0.013
#> 81        cognitive control -0.013
#> 112         decision making -0.013
#> 177                  eating -0.013
#> 224           eye movements -0.013
#> 337        major depression -0.013
#> 363              metabolism -0.013
#> 389         motor responses -0.013
#> 436                     ocd -0.013
#> 465             personality -0.013
#> 494         psychopathology -0.013
#> 556                 saccade -0.013
#> 561                 salient -0.013
#> 607                  sounds -0.013
#> 659                   trait -0.013
#> 674                upcoming -0.013
#> 689           visual motion -0.013
#> 51         auditory stimuli -0.014
#> 85                    color -0.014
#> 138              directions -0.014
#> 165                dopamine -0.014
#> 175            early visual -0.014
#> 210     executive functions -0.014
#> 434               obsessive -0.014
#> 446               orienting -0.014
#> 492             psychiatric -0.014
#> 48      attentional control -0.015
#> 59                    avoid -0.015
#> 136                   digit -0.015
#> 151            disorder ocd -0.015
#> 286             impulsivity -0.015
#> 323                 lesions -0.015
#> 370      monetary incentive -0.015
#> 557                saccades -0.015
#> 626                  stress -0.015
#> 662                  trauma -0.015
#> 9                adaptation -0.016
#> 14              adolescence -0.016
#> 39               arithmetic -0.016
#> 93               compulsive -0.016
#> 172                   drugs -0.016
#> 221               eye field -0.016
#> 495               psychosis -0.016
#> 630             stroop task -0.016
#> 133              developing -0.017
#> 189       emotional stimuli -0.017
#> 361                     met -0.017
#> 408      negative emotional -0.017
#> 596                 smokers -0.017
#> 612               spatially -0.017
#> 629                  stroop -0.017
#> 50                 auditory -0.018
#> 220                     eye -0.018
#> 322                  lesion -0.018
#> 562           schizophrenia -0.018
#> 570        self referential -0.018
#> 236           fearful faces -0.019
#> 240                  female -0.019
#> 609                 spatial -0.019
#> 37        anxiety disorders -0.020
#> 476                 playing -0.020
#> 702                 wm task -0.020
#> 179               education -0.021
#> 549                    risk -0.021
#> 633               substance -0.021
#> 36                  anxiety -0.023
#> 43                attention -0.023
#> 47              attentional -0.023
#> 124               depressed -0.023
#> 362               metabolic -0.023
#> 610       spatial attention -0.023
#> 15               adolescent -0.024
#> 376                    mood -0.024
#> 346                     mdd -0.025
#> 150            disorder mdd -0.026
#> 423                   noise -0.026
#> 127     depressive disorder -0.028
#> 161                     dmn -0.028
#> 621             spontaneous -0.028
#> 338        major depressive -0.029
#> 125              depression -0.030
#> 126              depressive -0.030
#> 378                  motion -0.030
#> 524            relationship -0.030
#> 413             network dmn -0.034
#> 535                 resting -0.035
#> 536           resting state -0.036